+++ Projects/STM32H743I-EVAL/Applications/LwIP/LwIP_HTTP_Server_Netconn_RTOS/Inc/main.h
@@ -6,7 +6,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2017-2021 STMicroelectronics.
+  * Copyright (c) 2017 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -26,31 +26,29 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32h743i_eval.h"
-#include "stm32h743i_eval_lcd.h"
 #include "stm32h743i_eval_sdram.h"
-#include "stm32_lcd.h"
 
 /* Exported types ------------------------------------------------------------*/
 /* Exported constants --------------------------------------------------------*/
-#define USE_LCD
+#define USE_LCD        /* enable LCD  */
 
 /*Static IP ADDRESS*/
-#define IP_ADDR0   192
-#define IP_ADDR1   168
-#define IP_ADDR2   0
-#define IP_ADDR3   10
+#define IP_ADDR0   ((uint8_t)192U)
+#define IP_ADDR1   ((uint8_t)168U)
+#define IP_ADDR2   ((uint8_t)0U)
+#define IP_ADDR3   ((uint8_t)10U)
 
 /*NETMASK*/
-#define NETMASK_ADDR0   255
-#define NETMASK_ADDR1   255
-#define NETMASK_ADDR2   255
-#define NETMASK_ADDR3   0
+#define NETMASK_ADDR0   ((uint8_t)255U)
+#define NETMASK_ADDR1   ((uint8_t)255U)
+#define NETMASK_ADDR2   ((uint8_t)255U)
+#define NETMASK_ADDR3   ((uint8_t)0U)
 
 /*Gateway Address*/
-#define GW_ADDR0   192
-#define GW_ADDR1   168
-#define GW_ADDR2   0
-#define GW_ADDR3   1
+#define GW_ADDR0   ((uint8_t)192U)
+#define GW_ADDR1   ((uint8_t)168U)
+#define GW_ADDR2   ((uint8_t)0U)
+#define GW_ADDR3   ((uint8_t)1U)
 
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions ------------------------------------------------------- */
